+++ src/serial.rs
@@ -25,6 +25,8 @@ pub enum Event {
     Rxne,
     /// New data can be sent
     Txe,
+    /// Idle line state detected
+    Idle,
 }
 
 /// Serial error
@@ -179,16 +181,18 @@ where
     /// Starts listening for an interrupt event
     pub fn listen(&mut self, event: Event) {
         match event {
-            Event::Rxne => self.usart.cr1.modify(|_, w| w.rxneie().set_bit()),
-            Event::Txe => self.usart.cr1.modify(|_, w| w.txeie().set_bit()),
+            Event::Rxne => self.usart.cr1.modify(|_, w| w.rxneie().enabled()),
+            Event::Txe => self.usart.cr1.modify(|_, w| w.txeie().enabled()),
+            Event::Idle => self.usart.cr1.modify(|_, w| w.idleie().enabled())
         }
     }
 
     /// Starts listening for an interrupt event
     pub fn unlisten(&mut self, event: Event) {
         match event {
-            Event::Rxne => self.usart.cr1.modify(|_, w| w.rxneie().clear_bit()),
-            Event::Txe => self.usart.cr1.modify(|_, w| w.txeie().clear_bit()),
+            Event::Rxne => self.usart.cr1.modify(|_, w| w.rxneie().disabled()),
+            Event::Txe => self.usart.cr1.modify(|_, w| w.txeie().disabled()),
+            Event::Idle => self.usart.cr1.modify(|_, w| w.idleie().disabled())
         }
     }